Refactor changed constants object from list to dictionary so that i can use keys for clarity instead of magic numbers
