+++ tests/background_detection_test.py
@@ -3,10 +3,9 @@ from unittest.mock import patch
 from uuid import UUID
 
 import pytest
+from api.models import Detection
+from api.tasks import background_detection
 from celery.exceptions import Retry
-
-from backend.api.models import Detection
-from backend.api.tasks import background_detection
 
 
 # Happy path tests
